Private types and functions

Documentation for OceanTurbulenceParameterEstimation.jl's internal interface.

OceanTurbulenceParameterEstimation

Observations

OceanTurbulenceParameterEstimation.Observations.FieldTimeSeriesCollector — Method

FieldTimeSeriesCollector(collected_fields, times; architecture=CPU())

Returns a FieldTimeSeriesCollector for fields of simulation. fields is a NamedTuple of AbstractFields that are to be collected.

OceanTurbulenceParameterEstimation.Observations.SyntheticObservations — Type

SyntheticObservations(path; field_names,
                      transformation = Transformation()),
                      times = nothing,
                      field_time_serieses = nothing,
                      regrid_size = nothing)

Return a time series of synthetic observations generated by Oceananigans.jl's simulations gridded as Oceananigans.jl fields.

OceanTurbulenceParameterEstimation.Observations.column_ensemble_interior — Method

column_ensemble_interior(observations::Vector{<:SyntheticObservations}, field_name, time_indices::Vector, N_ens)

Returns an Nensemble × Nbatch × Nz Array of (1, 1, Nz) field_name data, given Nbatch SyntheticObservations objects. The Nbatch × Nz data for field_name is copied Nensemble times to form a 3D Array.

OceanTurbulenceParameterEstimation.Observations.observation_names — Method

observation_names(obs::Vector{<:SyntheticObservations})

Return a Set representing the union of all names in obs.

Parameters

OceanTurbulenceParameterEstimation.Parameters.FreeParameters — Method

FreeParameters(priors; names = Symbol.(keys(priors)))

Return named FreeParameters with priors. Free parameter names are inferred from the keys of priors if not provided.

Example

julia> using Distributions, OceanTurbulenceParameterEstimation

julia> priors = (ν = Normal(1e-4, 1e-5), κ = Normal(1e-3, 1e-5))
(ν = Normal{Float64}(μ=0.0001, σ=1.0e-5), κ = Normal{Float64}(μ=0.001, σ=1.0e-5))

julia> free_parameters = FreeParameters(priors)
FreeParameters with 2 parameters
├── names: (:ν, :κ)
└── priors: Dict{Symbol, Any}
    ├── ν => Normal{Float64}(μ=0.0001, σ=1.0e-5)
    └── κ => Normal{Float64}(μ=0.001, σ=1.0e-5)

OceanTurbulenceParameterEstimation.Parameters.ScaledLogitNormal — Type

ScaledLogitNormal([FT=Float64;] bounds=(0, 1), mass=0.5, interval=nothing)

Return a ScaledLogitNormal distribution with compact support within bounds.

interval is an optional 2-element tuple or Array. When specified, the parameters μ and σ of the underlying Normal distribution are calculated so that mass fraction of the probability density lies within interval.

If interval is not specified, then μ=0 and σ=1 by default.

Notes

ScaledLogitNormal is a four-parameter distribution generated by the transformation

\[Y = L + (U - L) / [1 + \exp(X)],\]

of the normally-distributed variate $X ∼ 𝒩(μ, σ)$. The four parameters governing the distribution of $Y$ are thus

$L$: lower bound (0 for the LogitNormal distribution)

$U$: upper bound (1 for the LogitNormal distribution)

$μ$: mean of the underlying Normal distribution

$σ²$: variance of the underlying Normal distribution

OceanTurbulenceParameterEstimation.Parameters.closure_with_parameters — Method

closure_with_parameters(closure, parameters)

Returns a new object where for each (parameter_name, parameter_value) pair in parameters, the value corresponding to the key in object that matches parameter_name is replaced with parameter_value.

Example

julia> using OceanTurbulenceParameterEstimation.Parameters: closure_with_parameters

julia> struct ClosureSubModel; a; b end

julia> struct Closure; test; c end

julia> closure = Closure(ClosureSubModel(1, 2), 3)
Closure(ClosureSubModel(1, 2), 3)

julia> parameters = (a = 12, d = 7)
(a = 12, d = 7)

julia> closure_with_parameters(closure, parameters)
Closure(ClosureSubModel(12, 2), 3)

OceanTurbulenceParameterEstimation.Parameters.lognormal — Method

lognormal(; mean, std)

Return Lognormal distribution parameterized by the distribution mean and standard deviation std.

Notes

A variate X is LogNormal distributed if

\[\log(X) ∼ 𝒩(μ, σ²) ,\]

where $𝒩(μ, σ²)$ is the Normal distribution with mean $μ$ and variance $σ²$.

The mean and variance $s²$ (where $s$ is the standard deviation or std) are related to the parameters $μ$ and $σ²$ via

\[ m = \exp(μ + σ² / 2),\]

\[s² = [\exp(σ²) - 1] m².\]

These formula allow us to calculate $μ$ and $σ$ given $m$ and $s²$, since rearranging the formula for $s²$ gives

\[\exp(σ²) = m² / s² + 1\]

which then yields

\[σ = \sqrt{\log(m² / s² + 1)}.\]

We then find that

\[μ = \log(m) - σ² / 2 .\]

See also wikipedia.

OceanTurbulenceParameterEstimation.Parameters.normal_to_scaled_logit_normal — Method

Return a logit-normally distributed variate given the normally-distributed variate X.

OceanTurbulenceParameterEstimation.Parameters.scaled_logit_normal_to_normal — Method

Return a normally-distributed variate given the logit-normally distributed variate Y.

OceanTurbulenceParameterEstimation.Parameters.transform_to_constrained — Method

transform_to_constrained(Π, X)

Transform an "unconstrained", normally-distributed variate X to "constrained" (physical) space via the map associated with the distribution Π of Y.

OceanTurbulenceParameterEstimation.Parameters.transform_to_unconstrained — Method

transform_to_unconstrained(Π, Y)

Transform the "constrained" (physical) variate Y into it's unconstrained (normally-distributed) counterpart X through the forward map associated with Π.

If some mapping between $Y$ and the normally-distributed $X$ is defined via

\[Y = g(X).\]

Then transform_to_unconstrained is the inverse $X = g^{-1}(Y)$. The change of variables $g(X)$ determines the distribution Π of Y.

Example

The logarithm of a LogNormal(μ, σ) distributed variate is normally-distributed, such that the forward trasform $f ≡ \exp$,

\[Y = \exp(X),\]

and the inverse trasnform is the natural logarithm $f^{-1} ≡ \log$,

\[\log(Y) = X ∼ 𝒩(μ, σ).\]

OceanTurbulenceParameterEstimation.Parameters.update_closure_ensemble_member! — Method

update_closure_ensemble_member!(closures, p_ensemble, parameters)

Use parameters to update closure from closures that corresponds to ensemble member p_ensemble.

EnsembleKalmanInversions

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.EnsembleKalmanInversion — Method

EnsembleKalmanInversion(inverse_problem; noise_covariance=1e-2, resampler=Resampler())

Return an object that interfaces with EnsembleKalmanProcesses.jl and uses Ensemble Kalman Inversion to iteratively "solve" the inverse problem:

\[y = G(θ) + η,\]

for the parameters $θ$, where $y$ is a "normalized" vector of observations, $G(θ)$ is a forward map that predicts the observations, and $η ∼ 𝒩(0, Γ_y)$ is zero-mean random noise with covariance matrix $Γ_y$ representing uncertainty in the observations.

By "solve", we mean that the iteration finds the parameter values $θ$ that minimizes the distance between $y$ and $G(θ)$.

The "forward map output" G can have many interpretations. The specific statistics that G computes have to be selected for each use case to provide a concise summary of the complex model solution that contains the values that we would most like to match to the corresponding truth values y. For example, in the context of an ocean-surface boundary layer parametrization, this summary could be a vector of concatenated u, v, b, e profiles at all or some time steps of the CATKE solution.

(For more details on the Ensemble Kalman Inversion algorithm refer to the EnsembleKalmanProcesses.jl Documentation.)

Arguments

inverse_problem :: InverseProblem: Represents an inverse problem representing the comparison between synthetic observations generated by Oceananigans.jl and model predictions, also generated by Oceananigans.jl.

noise_covariance (AbstractMatrix or Number): normalized covariance representing observational uncertainty. If noise_covariance isa Number then it's converted to an identity matrix scaled by noise_covariance.

resampler: controls particle resampling procedure. See Resampler.

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.IterationSummary — Type

struct IterationSummary{P, M, C, V, E}

Container with information about each iteration of the Ensemble Kalman Process.

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.IterationSummary — Type

IterationSummary(eki, parameters, forward_map_output=nothing)

Return the summary for ensemble Kalman inversion eki with free parameters and forward_map_output.

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.column_has_nan — Method

Return a BitVector indicating which particles are NaN.

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.iterate! — Method

iterate!(eki::EnsembleKalmanInversion; iterations = 1, show_progress = true)

Iterate the ensemble Kalman inversion problem eki forward by iterations.

Return

best_parameters: the ensemble mean of all parameter values after the last iteration.

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.resample! — Method

resample!(resampler::Resampler, θ, G, eki)

Resamples the parameters θ of the eki process based on the number of NaN values inside the forward map output G.

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.sample — Method

sample(eki, θ, G, Nsample)

Generate Nsample new particles sampled from a multivariate Normal distribution parameterized by the ensemble mean and covariance computed based on the Nθ × Nensemble ensemble array θ, under the condition that all Nsample particles produce successful forward map outputs (don't include NaNs).

G (size(G) = Noutput × Nensemble) is the forward map output produced by θ.

Returns Nθ × Nsample parameter Array and Noutput × Nsample forward map output Array.