Private types and functions

Documentation for OceanTurbulenceParameterEstimation.jl's internal interface.

OceanTurbulenceParameterEstimation

Observations

OceanTurbulenceParameterEstimation.Observations.FieldTimeSeriesCollector — Method

FieldTimeSeriesCollector(collected_fields, times; architecture=CPU())

Returns a FieldTimeSeriesCollector for fields of simulation. fields is a NamedTuple of AbstractFields that are to be collected.

OceanTurbulenceParameterEstimation.Observations.SyntheticObservations — Type

SyntheticObservations{F, G, T, P, M} <: AbstractObservation

A time series of synthetic observations generated by Oceananigans.jl's simulations gridded as Oceananigans.jl fields.

OceanTurbulenceParameterEstimation.Observations.column_ensemble_interior — Method

column_ensemble_interior(observations::Vector{<:SyntheticObservations}, field_name, time_indices::Vector, N_ens)

Returns an N_cases × N_ens × Nz array of the interior of a field field_name defined on a OneDimensionalEnsembleGrid of size N_cases × N_ens × Nz, given a list of SyntheticObservations objects containing the N_cases single-column fields at time index in time_index.

OceanTurbulenceParameterEstimation.Observations.observation_names — Method

observation_names(ts_vector::Vector{<:SyntheticObservations})

Return a Set representing the union of all names in ts_vector.

TurbulenceClosureParameters

OceanTurbulenceParameterEstimation.TurbulenceClosureParameters.FreeParameters — Method

FreeParameters(priors; names = Symbol.(keys(priors)))

Return named FreeParameters with priors. Free parameter names are inferred from the keys of priors if not provided.

Example

julia> using Distributions, OceanTurbulenceParameterEstimation

julia> priors = (ν = Normal(1e-4, 1e-5), κ = Normal(1e-3, 1e-5))
(ν = Normal{Float64}(μ=0.0001, σ=1.0e-5), κ = Normal{Float64}(μ=0.001, σ=1.0e-5))

julia> free_parameters = FreeParameters(priors)
FreeParameters with 2 parameters
├── names: (:ν, :κ)
└── priors: Dict{Symbol, Any}
    ├── ν => Normal{Float64}(μ=0.0001, σ=1.0e-5)
    └── κ => Normal{Float64}(μ=0.001, σ=1.0e-5)

OceanTurbulenceParameterEstimation.TurbulenceClosureParameters.closure_with_parameters — Method

closure_with_parameters(closure, parameters)

Returns a new object where for each (parameter_name, parameter_value) pair in parameters, the value corresponding to the key in object that matches parameter_name is replaced with parameter_value.

Example

julia> using OceanTurbulenceParameterEstimation.TurbulenceClosureParameters: closure_with_parameters

julia> struct ClosureSubModel; a; b end

julia> struct Closure; test; c end

julia> closure = Closure(ClosureSubModel(1, 2), 3)
Closure(ClosureSubModel(1, 2), 3)

julia> parameters = (a = 12, d = 7)
(a = 12, d = 7)

julia> closure_with_parameters(closure, parameters)
Closure(ClosureSubModel(12, 2), 3)

OceanTurbulenceParameterEstimation.TurbulenceClosureParameters.update_closure_ensemble_member! — Method

update_closure_ensemble_member!(closures, p_ensemble, parameters)

Use parameters to update closure from closures that corresponds to ensemble member p_ensemble.

EnsembleKalmanInversions

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.EnsembleKalmanInversion — Method

EnsembleKalmanInversion(inverse_problem; noise_covariance=1e-2, resampler=NaNResampler())

Return an object that interfaces with EnsembleKalmanProcesses.jl and uses Ensemble Kalman Inversion to iteratively "solve" the inverse problem:

\[y = G(θ) + η,\]

for the parameters $θ$, where $y$ is a "normalized" vector of observations, $G(θ)$ is a forward map that predicts the observations, and $η ∼ N(0, Γ_y)$ is zero-mean random noise with covariance matrix $Γ_y$ representing uncertainty in the observations.

By "solve", we mean that the iteration finds the parameter values $θ$ that minimizes the distance between $y$ and $G(θ)$.

The "forward map output" G can have many interpretations. The specific statistics that G computes have to be selected for each use case to provide a concise summary of the complex model solution that contains the values that we would most like to match to the corresponding truth values y. For example, in the context of an ocean-surface boundary layer parametrization, this summary could be a vector of concatenated u, v, b, e profiles at all or some time steps of the CATKE solution.

(For more details on the Ensemble Kalman Inversion algorithm refer to the EnsembleKalmanProcesses.jl Documentation.)

Arguments

inverse_problem :: InverseProblem: Represents an inverse problem representing the comparison between synthetic observations generated by Oceananigans.jl and model predictions, also generated by Oceananigans.jl.

noise_covariance (AbstractMatrix or Number): normalized covariance representing observational uncertainty. If noise_covariance isa Number then it's converted to an identity matrix scaled by noise_covariance.

resampler: controls resampling procedure when the forward map contains NaNs. See NaNResampler.

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.IterationSummary — Type

struct IterationSummary{P, M, C, V, E}

Container with information about each iteration of the Ensemble Kalman Process.

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.IterationSummary — Type

IterationSummary(eki, parameters, forward_map_output=nothing)

Return the summary for ensemble Kalman inversion eki with free parameters and forward_map_output.

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.iterate! — Method

iterate!(eki::EnsembleKalmanInversion; iterations=1)

Iterate the ensemble Kalman inversion problem eki forward by iterations.

Return

best_parameters: the ensemble mean of all parameter values after the last iteration.

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.parameter_ensemble — Method

parameter_ensemble(eki::EnsembleKalmanInversion)

Return a Vector of parameter sets (in physical / constrained space) for each ensemble member.

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.resample! — Method

resample!(resampler::NaNResampler, G, θ, eki)

Resamples the parameters θ of the eki process based on the number of NaN values inside the forward map output G.

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.sample — Method

sample(eki, θ, G, n)

Generate n new particles sampled from a multivariate Normal distribution parameterized by the ensemble mean and covariance computed based on the N_θ × N_ensemble ensemble array θ, under the condition that all n particles lead to forward map outputs that are "stable" (don't include NaNs). G is the inverting forward map computed on ensemble θ.

Return an N_θ × n array of new particles, along with the inverting forward map output corresponding to the new particles.