Private types and functions

Documentation for OceanTurbulenceParameterEstimation.jl's internal interface.

OceanTurbulenceParameterEstimation

Observations

OceanTurbulenceParameterEstimation.Observations.FieldTimeSeriesCollector — Method

FieldTimeSeriesCollector(fields)

Returns a FieldTimeSeriesCollector for fields of simulation.

fields is a NamedTuple of AbstractFields that are to be collected.

OceanTurbulenceParameterEstimation.Observations.OneDimensionalTimeSeries — Type

OneDimensionalTimeSeries{F, G, T, P, M} <: AbstractObservation

A time series of horizontally-averaged observational or LES data gridded as Oceananigans fields.

OceanTurbulenceParameterEstimation.Observations.column_ensemble_interior — Method

column_ensemble_interior(observations::Vector{<:OneDimensionalTimeSeries}, field_name, time_indices::Vector, N_ens)

Returns an N_cases × N_ens array of the interior of a field field_name defined on a OneDimensionalEnsembleGrid of size N_cases × N_ens × Nz, given a list of OneDimensionalTimeSeries objects containing the N_cases single-column fields at the corresponding time index in time_indices.

TurbulenceClosureParameters

OceanTurbulenceParameterEstimation.TurbulenceClosureParameters.FreeParameters — Method

FreeParameters(priors; names = Symbol.(keys(priors)))

Return named FreeParameters with priors. Free parameter names are inferred from the keys of priors if not provided.

Examples

julia> priors = (ν = Normal(1e-4, 1e-5), κ = Normal(1e-3, 1e-5))

julia> free_parameters = FreeParameter(priors)

OceanTurbulenceParameterEstimation.TurbulenceClosureParameters.closure_with_parameters — Method

closure_with_parameters(closure, parameters)

Returns a new object where for each (parameter_name, parameter_value) pair in parameters, the value corresponding to the key in object that matches parameter_name is replaced with parameter_value.

Example

julia> struct ClosureSubModel a; b end

julia> struct Closure test; c end

julia> closure = Closure(ClosureSubModel(1, 2), 3)
Closure(ClosureSubModel(1, 2), 3)

julia> parameters = (a = 12, c = 7)
(a = 12, c = 7)

julia> closure_with_parameters(closure, parameters)
Closure(ClosureSubModel(12, 2), 7)

EnsembleKalmanInversions

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.EnsembleKalmanInversion — Method

EnsembleKalmanInversion(inverse_problem; noise_covariance=1e-2)

Return an object that interfaces with EnsembleKalmanProcesses.jl to iteratively "solve" the inverse problem

\[y = G(θ) + η,\]

for the parameters $θ$, where $y$ is a "normalized" vector of observations, $G(θ)$ is a forward map that predicts the observations, and $η ∼ N(0, Γy)$ is zero-mean random noise with covariance matrix $Γy$ representing uncertainty in the observations.

By "solve", we mean that the iteration finds $θ$ that minimizes the distance between $y$ and $G(θ)$.

The "forward map output" G can have many interpretations. The specific statistics that G computes have to be selected for each use case to provide a concise summary of the complex model solution that contains the values that we would most like to match to the corresponding truth values y. In the OSBL context, this summary could be, for example, a vector of concatenated u, v, b, e profiles at all or some time steps of the CATKE solution.

Arguments

inverse_problem::InverseProblem: Represents an inverse problem representing the comparison between synthetic observations generated by Oceananigans and model predictions also generated by Oceananigans.

noise_covariance (AbstractMatrix or Number): normalized covariance representing observational uncertainty. If noise_covariance isa Number then it's converted to an identity matrix scaled by noise_covariance.

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.UnscentedKalmanInversion — Method

UnscentedKalmanInversion(inverse_problem, prior_mean, prior_cov; noise_covariance=1e-2, α_reg = 1.0, update_freq = 0)

Arguments

inverse_problem : inverse problem

prior_mean::Array{Float64} : prior mean

prior_cov::Array{Float64, 2} : prior covariance

noise_covariance::Float64 : observation error covariance

α_reg::Float64 : regularization parameter toward u0 (0 < α_reg ≤ 1), default should be 1, without regulariazion

update_freq::IT : set to 0 when the inverse problem is not identifiable (default), namely the inverse problem has multiple solutions, the covariance matrix will represent only the sensitivity of the parameters, instead of posterior covariance information; set to 1 (or anything > 0) when the inverse problem is identifiable, and the covariance matrix will converge to a good approximation of the posterior covariance with an uninformative prior.

OceanTurbulenceParameterEstimation.EnsembleKalmanInversions.iterate! — Method

iterate!(eki::EnsembleKalmanInversion; iterations=1)

Iterate the ensemble Kalman inversion problem eki forward by iterations.